[#601] [agent:backend-implementer] harden self-review: int8 accrual, register graph_id validation, best-effort accrue
- recurring_cost_tokens: Integer -> BigInteger (int8) — a lifetime accumulator on a
  truly-standing team would overflow int4 (model + migration 0015).
- register(target_kind=team) now fail-closes a cross-org / non-existent graph_id via a
  KGS existence check (ScheduleService._validate_graph_id, GraphClient wired through DI) —
  mirrors the request path (#524), so a bad binding is rejected at register, not per fire.
- _accrue_schedule_cost is best-effort (contextlib.suppress): the run is already settled
  terminal at the call site, so a transient accrual error must not error the worker task.

+++ services/execution-engine-service/migrations/versions/0015_standing_team_binding.py
@@ -34,7 +34,8 @@ def upgrade() -> None:
     op.add_column("engine_schedules", sa.Column("graph_id", sa.Text(), nullable=True))
     op.add_column(
         "engine_schedules",
-        sa.Column("recurring_cost_tokens", sa.Integer(), nullable=False, server_default="0"),
+        # BigInteger (int8): a lifetime accumulator on a truly-standing team would overflow int4.
+        sa.Column("recurring_cost_tokens", sa.BigInteger(), nullable=False, server_default="0"),
     )
     op.add_column(
         "engine_team_runs",
